+++ tests/clients/test_cases.py
@@ -38,11 +38,11 @@ class TestCasesClient:
         cases_client.get_all()
         mock_cases_service.get_all.call_count == 1
 
-    def test_get_case_by_case_number_calls_cases_service_with_expected_params(
+    def test_get_case_calls_cases_service_with_expected_params(
         self, mock_cases_service, mock_cases_file_event_service
     ):
         cases_client = CasesClient(mock_cases_service, mock_cases_file_event_service)
-        cases_client.get_case_by_case_number(123456)
+        cases_client.get_case(123456)
         mock_cases_service.get_case.assert_called_once_with(123456)
 
     def test_export_calls_cases_service_with_expected_params(