+++ CHANGELOG.md
@@ -22,10 +22,10 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
     - `sdk.cases.create()`
     - `sdk.cases.get_all()`
     - `sdk.cases.export_summary()`
-    - `sdk.cases.get_case_by_case_number()`
+    - `sdk.cases.get_case()`
     - `sdk.case.update()`
 
-- `sdk.cases.file_events` methods for managing cases file events:
+- `sdk.cases.file_events` methods for managing file events associated with a given case:
     - `sdk.cases.file_events.add_event`
     - `sdk.cases.file_events.delete_event`
     - `sdk.cases.file_events.get_event`